add proses approval

+++ resources/views/pages/s_user/karyawan/m_rembes_item/index.blade.php
@@ -31,17 +31,19 @@
 
 
             @can('rembes-item-create')
-                <a class="btn btn-primary" href="{{ route('dashboard.rembes-item.create', $rembes->id) }}">
-                    <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none"
-                        stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"
-                        class="feather feather-user-plus">
-                        <path d="M16 21v-2a4 4 0 0 0-4-4H5a4 4 0 0 0-4 4v2"></path>
-                        <circle cx="8.5" cy="7" r="4"></circle>
-                        <line x1="20" y1="8" x2="20" y2="14"></line>
-                        <line x1="23" y1="11" x2="17" y2="11"></line>
-                    </svg>
-                    Add New Reimburse Item
-                </a>
+                @if ($rembes->status !== 'APPROVED')
+                    <a class="btn btn-primary" href="{{ route('dashboard.rembes-item.create', $rembes->id) }}">
+                        <svg xmlns="http://www.w3.org/2000/svg" width="24" height="24" viewBox="0 0 24 24" fill="none"
+                            stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"
+                            class="feather feather-user-plus">
+                            <path d="M16 21v-2a4 4 0 0 0-4-4H5a4 4 0 0 0-4 4v2"></path>
+                            <circle cx="8.5" cy="7" r="4"></circle>
+                            <line x1="20" y1="8" x2="20" y2="14"></line>
+                            <line x1="23" y1="11" x2="17" y2="11"></line>
+                        </svg>
+                        Add New Reimburse Item
+                    </a>
+                @endif
             @endcan
         </div>
 
@@ -101,27 +103,35 @@
                                     </td>
 
                                     <td>
-
+                                        <a class="badge badge-light-primary text-start me-2"
+                                            href="{{ route('dashboard.rembes-item.show', ['rembes' => $rembes->id, 'id' => $rembes_item->id]) }}">
+                                            <i class="far fa-edit"></i>
+                                            Detail
+                                        </a>
                                         @can('rembes-item-edit')
-                                            <a class="badge badge-light-primary text-start me-2"
-                                                href="{{ route('dashboard.rembes-item.edit', ['rembes' => $rembes->id, 'id' => $rembes_item->id]) }}">
-                                                <i class="far fa-edit"></i>
-                                                Edit
-                                            </a>
+                                            @if ($rembes->status !== 'APPROVED')
+                                                <a class="badge badge-light-primary text-start me-2"
+                                                    href="{{ route('dashboard.rembes-item.edit', ['rembes' => $rembes->id, 'id' => $rembes_item->id]) }}">
+                                                    <i class="far fa-edit"></i>
+                                                    Edit
+                                                </a>
+                                            @endif
                                         @endcan
 
                                         @can('rembes-item-delete')
-                                            {!! Form::open([
-                                                'method' => 'DELETE',
-                                                'route' => ['dashboard.rembes-item.destroy', ['rembes' => $rembes->id, 'id' => $rembes_item->id]],
-                                                'style' => 'display:inline',
-                                            ]) !!}
-                                            {!! Form::button('<i class="far fa-trash-alt"></i> Delete', [
-                                                'type' => 'submit',
-                                                'class' => 'badge badge-light-danger text-start mx-3',
-                                                'onclick' => 'return confirm("Are you sure you want to delete this rembes item?");',
-                                            ]) !!}
-                                            {!! Form::close() !!}
+                                            @if ($rembes->status !== 'APPROVED')
+                                                {!! Form::open([
+                                                    'method' => 'DELETE',
+                                                    'route' => ['dashboard.rembes-item.destroy', ['rembes' => $rembes->id, 'id' => $rembes_item->id]],
+                                                    'style' => 'display:inline',
+                                                ]) !!}
+                                                {!! Form::button('<i class="far fa-trash-alt"></i> Delete', [
+                                                    'type' => 'submit',
+                                                    'class' => 'badge badge-light-danger text-start mx-3',
+                                                    'onclick' => 'return confirm("Are you sure you want to delete this rembes item?");',
+                                                ]) !!}
+                                                {!! Form::close() !!}
+                                            @endif
                                         @endcan
                                     </td>
                                 </tr>
